Codebase clean up.
