+++ kornia/constants.py
@@ -1,6 +1,9 @@
 from typing import Union, TypeVar
 from enum import Enum
 
+import torch
+
+PI = torch.tensor(3.14159265358979323846)
 T = TypeVar('T', bound='Resample')
