+++ kornia/geometry/transform/crop.py
@@ -144,7 +144,8 @@ def center_crop(tensor: torch.Tensor, size: Tuple[int, int], interpolation: str 
         [dst_w - 1, dst_h - 1],
         [0, dst_h - 1],
     ]], device=tensor.device).expand(points_src.shape[0], -1, -1)
-    return crop_by_boxes(tensor, points_src, points_dst, interpolation, return_transform=return_transform)
+    return crop_by_boxes(
+        tensor, points_src.float(), points_dst.float(), interpolation, return_transform=return_transform)
 
 
 def crop_by_boxes(tensor, src_box, dst_box, interpolation: str = 'bilinear',
@@ -160,12 +161,9 @@ def crop_by_boxes(tensor, src_box, dst_box, interpolation: str = 'bilinear',
         tensor = torch.unsqueeze(tensor, dim=0)
 
     # compute transformation between points and warp
-    dst_trans_src: torch.Tensor = get_perspective_transform(
-        src_box.to(tensor.device).to(tensor.dtype),
-        dst_box.to(tensor.device).to(tensor.dtype)
-    )
+    dst_trans_src: torch.Tensor = get_perspective_transform(src_box, dst_box)
     # simulate broadcasting
-    dst_trans_src = dst_trans_src.expand(tensor.shape[0], -1, -1)
+    dst_trans_src = dst_trans_src.expand(tensor.shape[0], -1, -1).type_as(tensor)
 
     bbox = _infer_bounding_box(dst_box)
     patches: torch.Tensor = warp_affine(
